--- common/v2/providers/DerivationPathOptionsProvider/index.ts
@@ -1,1 +1,0 @@
-export * from './DerivationPathOptionsProvider';
