+++ common/v2/providers/index.ts
@@ -3,10 +3,6 @@ export { AccountContext, AccountProvider } from './AccountProvider';
 export { WalletContext, WalletProvider } from './WalletProvider';
 export { AssetContext, AssetProvider } from './AssetProvider';
 export { ContractsContext, ContractsProvider } from './ContractsProvider';
-export {
-  DerivationPathOptionsContext,
-  DerivationPathOptionsProvider
-} from './DerivationPathOptionsProvider';
 export { SettingsContext, SettingsProvider } from './SettingsProvider';
 export {
   NotificationsContext,
