--- common/v2/services/DerivationPathOptions/index.ts
@@ -1,2 +1,0 @@
-export * from './DerivationPathOptions';
-export * from './types';
