--- common/v2/services/DerivationPathOptions/types.ts
@@ -1,9 +1,0 @@
-export interface DerivationPathOptions {
-  name: string;
-  derivationPath: string;
-  active: boolean;
-}
-
-export interface ExtendedDerivationPathOptions extends DerivationPathOptions {
-  uuid: string;
-}
