+++ common/v2/services/types.ts
@@ -4,7 +4,6 @@ export * from 'v2/services/Notifications/types';
 export * from 'v2/services/AddressBook/types';
 export * from 'v2/services/Asset/types';
 export * from 'v2/services/Contracts/types';
-export * from 'v2/services/DerivationPathOptions/types';
 export * from 'v2/services/Settings/types';
 export * from 'v2/services/Network/types';
 export * from 'v2/services/ScreenLockSettings/types';
